+++ public/javascripts/index.js
@@ -1,10 +1,18 @@
-$(document).ready( function(){
+moment.locale('sv');
 
-  $('.clockpicker').clockpicker();
-  
-  if(document.getElementById('diaryform')) document.getElementById('diaryform').onsubmit = validateForm;
-
-});
+function delete_post(id, target_element_id, redirect = false){
+  var target_element = document.getElementById(target_element_id);
+  var xhr1 = new XMLHttpRequest();
+  xhr1.open('DELETE', "/posts/"+ id, true);
+  xhr1.onreadystatechange = function() {
+      if (this.status == 200 && this.readyState == 4) {
+          console.log(target_element);
+          if(target_element) target_element.className = target_element.className + "collapse";
+          if(redirect) window.location = "/"; 
+      }
+  };//end onreadystate
+  xhr1.send();
+}
 
 function times(count, callbackOrScalar) {
     let type = typeof callbackOrScalar
