+++ routes/users.js
@@ -1,8 +1,8 @@
-var express = require('express');
-var router = express.Router();
-var db = require('../db.js');
-var md5 = require('md5');
-var utils = require('../utils');
+const express = require('express');
+const router = express.Router();
+const db = require('../db.js');
+const md5 = require('md5');
+const helpers = require('../utils/helpers');
 
 /* NEW user form */
 router.get('/', function(req, res, next) {
@@ -13,7 +13,7 @@ router.get('/', function(req, res, next) {
 router.post('/', function(req, res, next) {
   var hash = md5(Date.now()).slice(0,12);
   db.run(`INSERT INTO users (hash) VALUES (?)`,[hash], function(err){
-    res.render('new_user', { link: utils.link_to(req, `${this.lastID}/${hash}`), hash: hash, title: "Ny dagbok" });
+    res.render('new_user', { link: helpers.link_to(req, `${this.lastID}/${hash}`), hash: hash, title: "Ny dagbok" });
   }) 
 });
