+++ views/_streaks.ejs
@@ -1,93 +1,36 @@
-<!--_streaks.ejs -->
-  <div class="card card-background">
-    <div class="card-body streak-card">
+<% 
+switch(notice.what){  
+case 'window-hit':
+%> 
 
-        <h2> Notiser </h2>
+    <p>Du prickade ditt sömnfönster i ditt senaste inlägg!</p>
 
-          <div id="streak-0" class="streak-partial ">
-
-              <p>
-                Du har loggat in i sömndagboken, ett steg i rätt riktning för att få bättre sömn!
-              </p>
-
-              <div class="row">
-                <svg class="medallion" width="61" height="55" viewBox="0 0 61 55" fill="none" xmlns="http://www.w3.org/2000/svg">
-                  <path d="M20.5968 36.8354L0 10.3981L13.3466 5.48162e-06L33.9434 26.4373L21.789 24.3201L20.5968 36.8354Z" fill="#6E90C4"/>
-                  <path d="M40.3465 36.8354L60.9434 10.3981L47.5968 5.48162e-06L26.9999 26.4373L39.1543 24.3201L40.3465 36.8354Z" fill="#6E90C4"/>
-                  <circle id="circle" cx="30.3813" cy="37.6164" r="17.3613" fill="#E0DAA0"/>
-                  <path d="M29.8308 25.5854C29.8906 25.4011 30.1513 25.4011 30.2112 25.5854L32.8995 33.8591C32.9262 33.9415 33.003 33.9973 33.0897 33.9973H41.7891C41.9828 33.9973 42.0634 34.2452 41.9067 34.3591L34.8687 39.4725C34.7986 39.5234 34.7692 39.6137 34.796 39.6961L37.4843 47.9697C37.5442 48.154 37.3333 48.3072 37.1765 48.1933L30.1385 43.0799C30.0684 43.029 29.9735 43.029 29.9034 43.0799L22.8654 48.1933C22.7087 48.3072 22.4978 48.154 22.5576 47.9697L25.2459 39.6961C25.2727 39.6137 25.2434 39.5234 25.1733 39.4725L18.1353 34.3591C17.9785 34.2452 18.0591 33.9973 18.2528 33.9973H26.9523C27.0389 33.9973 27.1157 33.9415 27.1425 33.8591L29.8308 25.5854Z" fill="#D3CC8A"/>
-                </svg>
-              </div>
-
-          </div>
-
-          <div id="streak-1" class="streak-partial collapse">
-
-              <p>
-                Du prickade ditt sömnfönster i ditt senaste inlägg!
-              </p>
-
-
-                <div class="smiley">
-                  <img src="images/smileyHappy.svg">
-                </div>
-
-
-          </div>
-
-          <div id="streak-2" class="streak-partial collapse">
-
-              <p>
-                Du missade tyvärr ditt sömnfönster i ditt senaste inlägg!
-              </p>
-
-              <div class="smiley">
-                  <img src="images/smileySad.svg">
-              </div>
-
-          </div>
-
-
-          <div class="row">
-            <div class="col">
-                <div class="mb-4">
-                  <a href="/posts/new" class="btn btn-primary btn-skapa btn-spara">
-                    <i class="fa fa-plus"></i> Registrera sömn 
-                  </a>
-                </div>
-              </div>
-          </div>
-
-              <!-- URSPRUNGSDESIGN
-              <div class="row">
-                <div class="col-8">
-                  <p class="veckans-tips">
-                    Du har fyllt i sömndagboken:
-                   </p>
-                   <p class="veckans-tips streak-text">
-                      <span class="streak"> 5 </span> dagar i rad.
-                  </p>
-                </div>
-
-                <div class="col">
-                  <svg class="medallion" width="61" height="55" viewBox="0 0 61 55" fill="none" xmlns="http://www.w3.org/2000/svg">
-                    <path d="M20.5968 36.8354L0 10.3981L13.3466 5.48162e-06L33.9434 26.4373L21.789 24.3201L20.5968 36.8354Z" fill="#6E90C4"/>
-                    <path d="M40.3465 36.8354L60.9434 10.3981L47.5968 5.48162e-06L26.9999 26.4373L39.1543 24.3201L40.3465 36.8354Z" fill="#6E90C4"/>
-                    <circle id="circle" cx="30.3813" cy="37.6164" r="17.3613" fill="#E0DAA0"/>
-                    <path d="M29.8308 25.5854C29.8906 25.4011 30.1513 25.4011 30.2112 25.5854L32.8995 33.8591C32.9262 33.9415 33.003 33.9973 33.0897 33.9973H41.7891C41.9828 33.9973 42.0634 34.2452 41.9067 34.3591L34.8687 39.4725C34.7986 39.5234 34.7692 39.6137 34.796 39.6961L37.4843 47.9697C37.5442 48.154 37.3333 48.3072 37.1765 48.1933L30.1385 43.0799C30.0684 43.029 29.9735 43.029 29.9034 43.0799L22.8654 48.1933C22.7087 48.3072 22.4978 48.154 22.5576 47.9697L25.2459 39.6961C25.2727 39.6137 25.2434 39.5234 25.1733 39.4725L18.1353 34.3591C17.9785 34.2452 18.0591 33.9973 18.2528 33.9973H26.9523C27.0389 33.9973 27.1157 33.9415 27.1425 33.8591L29.8308 25.5854Z" fill="#D3CC8A"/>
-                  </svg>
-                </div>
-
-              </div>
-
-            <div class="row">
-              <div class="col">
-                  <div class="mb-4">
-                    <button class="btn btn-primary btn-spara" data-bs-toggle="collapse" data-bs-target="#form_row" ><i class="fa fa-plus"></i> Skapa nytt inlägg </button>
-                  </div>
-                </div>
-            </div>
-          -->
-
+    <div class="smiley">
+      <img src="images/smileyHappy.svg">
     </div>
+
+<% 
+break; 
+case 'window-miss':
+%>
+    <p>       Du missade tyvärr ditt sömnfönster i ditt senaste inlägg!       </p>
+
+    <div class="smiley">
+        <img src="images/smileySad.svg">
+    </div>
+<%
+break;
+default: 
+%>
+  <p>    Du har loggat in i sömndagboken - ett steg i riktning mot bättre sömn!  </p>
+
+  <div class="row">
+    <svg class="medallion" width="61" height="55" viewBox="0 0 61 55" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M20.5968 36.8354L0 10.3981L13.3466 5.48162e-06L33.9434 26.4373L21.789 24.3201L20.5968 36.8354Z" fill="#6E90C4"/>
+      <path d="M40.3465 36.8354L60.9434 10.3981L47.5968 5.48162e-06L26.9999 26.4373L39.1543 24.3201L40.3465 36.8354Z" fill="#6E90C4"/>
+      <circle id="circle" cx="30.3813" cy="37.6164" r="17.3613" fill="#E0DAA0"/>
+      <path d="M29.8308 25.5854C29.8906 25.4011 30.1513 25.4011 30.2112 25.5854L32.8995 33.8591C32.9262 33.9415 33.003 33.9973 33.0897 33.9973H41.7891C41.9828 33.9973 42.0634 34.2452 41.9067 34.3591L34.8687 39.4725C34.7986 39.5234 34.7692 39.6137 34.796 39.6961L37.4843 47.9697C37.5442 48.154 37.3333 48.3072 37.1765 48.1933L30.1385 43.0799C30.0684 43.029 29.9735 43.029 29.9034 43.0799L22.8654 48.1933C22.7087 48.3072 22.4978 48.154 22.5576 47.9697L25.2459 39.6961C25.2727 39.6137 25.2434 39.5234 25.1733 39.4725L18.1353 34.3591C17.9785 34.2452 18.0591 33.9973 18.2528 33.9973H26.9523C27.0389 33.9973 27.1157 33.9415 27.1425 33.8591L29.8308 25.5854Z" fill="#D3CC8A"/>
+    </svg>
   </div>
+
+<% } %>
