+++ views/index.ejs
@@ -19,7 +19,7 @@
       </div>
     </div>
 
-    <!--                      Översta raden                   -->
+    <!--                      Översta raden: Sömnfönster | Notiser                   -->
     <div class="row">
 
       <!-- Veckans Tips and Sleepview-->
@@ -31,9 +31,25 @@
 
       <!-- Streaks-->
       <div class="col-sm">
-        <%- include('_streaks') %>
-      </div>
+        <div class="card card-background">
+          <div class="card-body streak-card">
 
+            <h2> Notiser </h2>
+
+            <%- include('_streaks') %>
+            <div class="row">
+              <div class="col">
+                  <div class="mb-4">
+                    <a href="/posts/new" class="btn btn-primary btn-skapa btn-spara">
+                      <i class="fa fa-plus"></i> Registrera sömn 
+                    </a>
+                  </div>
+                </div>
+            </div>
+
+          </div>
+        </div>
+      </div>
     </div>
 
 
@@ -41,28 +57,18 @@
     <div class="row">
       <!-- Veckans Inlägg -->
       <div class="col" id="col-for-posts">
-      <%- include('_veckans_inlagg') %>
+        <%- include('_veckans_inlagg') %>
       </div>
-    </div> <!-- row -->
+    </div> 
+
+
+    <div class="btn-see">
+      <button class="btn btn-primary btn-avbryt" data-bs-toggle="collapse" data-bs-target="#all_posts_div"> Se mer </button>
+    </div>
 
     <div class="row collapse" id="all_posts_div">
       <div class="col">
-        <table id="all_posts_table" class="table">
-          <legend>Samtliga poster</legend>
-          <thead>
-            <th>Vecka</th>
-            <th>Sänggång</th>
-            <th>Klev upp</th>
-            <th>Vakentid</th>
-            <th>Sömneffektivitet</th>
-            <th>Sovtid</th>
-            <th>Självskattning</th>
-            <th>Registrerat</th>
-            <th>Radera</th>
-          </thead>
-          <tbody>
-          </tbody>
-        </table>
+        <%- include('_posts_table') %>
       </div>
     </div>
